+++ src/irmin-unix/resolver.mli
@@ -20,6 +20,19 @@ val global_option_section : string
 
 val branch : string option Cmdliner.Term.t
 
+(** {1 Hash} *)
+module Hash : sig
+  type t = (module Irmin.Hash.S)
+
+  val add : string -> ?default:bool -> (module Irmin.Hash.S) -> unit
+
+  val find : string -> t
+
+  val term : t option Cmdliner.Term.t
+end
+
+type hash = Hash.t
+
 (** {1 Contents} *)
 module Contents : sig
   type t = (module Irmin.Contents.S)
@@ -41,21 +54,27 @@ module Store : sig
      contains: the store implementation a creator of store's state and
      endpoint. *)
 
+  (** The type of constructors of a store configuration. Depending on the
+      backend, a store may require a hash function. *)
+  type store_functor =
+    | Fixed_hash of (contents -> t)
+    | Variable_hash of (hash -> contents -> t)
+
   type remote_fn = ?headers:Cohttp.Header.t -> string -> Irmin.remote
 
   val v : ?remote:remote_fn -> (module Irmin.S) -> t
 
-  val mem : contents -> t
+  val mem : hash -> contents -> t
 
-  val irf : contents -> t
+  val irf : hash -> contents -> t
 
   val http : t -> t
 
   val git : contents -> t
 
-  val find : string -> contents -> t
+  val find : string -> store_functor
 
-  val add : string -> ?default:bool -> (contents -> t) -> unit
+  val add : string -> ?default:bool -> store_functor -> unit
 end
 
 type Irmin.remote += R of Cohttp.Header.t option * string