+++ CHANGES.md
@@ -1,3 +1,23 @@
+### Unreleased
+
+#### Added
+
+- **irmin-unix**:
+  - Added a `--hash` parameter to the command-line interface, allowing the hash
+    function to be specified. For BLAKE2b and BLAKE2s, the bit-length may be
+    specified with a trailing slash, as in `--hash=blake2b/16`. The `hash`
+    function may also be specified in the configuration file. (#898, @craigfe)
+ 
+- **irmin**:
+  - Added `Irmin.Hash.Make_BLAKE2B` and `Irmin.Hash.Make_BLAKE2S` functors for
+    customizing the bit-length of these hash functions. (#898, @craigfe)
+
+#### Changed
+
+- **irmin-pack**:
+  - Changed the bit-length of serialized hashes from 60 to 30. (#897,
+    @icristescu)
+
 ### 2.0.0
 
 #### Added